refactor setup
